Build cleanup (#149)
Clean up build, move to L4 warnings and fix the ones that exist.
#121

+++ trlevel/ILevel.h
@@ -111,7 +111,7 @@ namespace trlevel
         // Get the mesh referenced by the specified mesh pointer.
         // mesh_pointer: The mesh pointer index.
         // Returns: The mesh.
-        virtual tr_mesh get_mesh_by_pointer(uint16_t mesh_pointer) const = 0;
+        virtual tr_mesh get_mesh_by_pointer(uint32_t mesh_pointer) const = 0;
 
         // Get the mesh tree node at the specified index.
         // index: The starting mesh tree index.
@@ -133,7 +133,7 @@ namespace trlevel
         // sprite_sequence_id: The id of the sprite sequence to find.
         // sequence: The place to store the sequence.
         // Returns: Whether the sprite sequence was found.
-        virtual bool get_sprite_sequence_by_id(uint32_t sprite_sequence_id, tr_sprite_sequence& sequence) const = 0;
+        virtual bool get_sprite_sequence_by_id(int32_t sprite_sequence_id, tr_sprite_sequence& sequence) const = 0;
 
         // Get the sprite texture with the specified ID.
         // index: The index of the sprite texture to get.
