+++ trlevel/Level.h
@@ -117,7 +117,7 @@ namespace trlevel
         // Get the mesh at the specified index.
         // index: The index of the mesh to get.
         // Returns: The mesh.
-        virtual tr_mesh get_mesh_by_pointer(uint16_t mesh_pointer) const override;
+        virtual tr_mesh get_mesh_by_pointer(uint32_t mesh_pointer) const override;
 
         // Get the mesh tree node at the specified index.
         // index: The mesh tree index.
@@ -139,7 +139,7 @@ namespace trlevel
         // sprite_sequence_id: The id of the sprite sequence to find.
         // sequence: The place to store the sequence.
         // Returns: Whether the sprite sequence was found.
-        virtual bool get_sprite_sequence_by_id(uint32_t sprite_sequence_id, tr_sprite_sequence& sequence) const override;
+        virtual bool get_sprite_sequence_by_id(int32_t sprite_sequence_id, tr_sprite_sequence& sequence) const override;
 
         // Get the sprite texture with the specified ID.
         // index: The index of the sprite texture to get.
